feat: Redesign bug reporting and header actions
- Remove BugReportTab component and move bug reporting to header
- Add bug report icon to main header and profile dropdown
- Add sync button to main header with deploy button styling
- Remove duplicate sync/bug report buttons from workbench header
- Clean up unused imports and code
- Improve header button organization and visibility
