+++ app/components/@settings/core/ControlPanel.tsx
@@ -26,7 +26,6 @@ import CloudProvidersTab from '~/components/@settings/tabs/providers/cloud/Cloud
 import ServiceStatusTab from '~/components/@settings/tabs/providers/status/ServiceStatusTab';
 import LocalProvidersTab from '~/components/@settings/tabs/providers/local/LocalProvidersTab';
 import McpTab from '~/components/@settings/tabs/mcp/McpTab';
-import BugReportTab from '~/components/@settings/tabs/bug-report/BugReportTab';
 
 interface ControlPanelProps {
   open: boolean;
@@ -143,8 +142,7 @@ export const ControlPanel = ({ open, onClose }: ControlPanelProps) => {
         return <ServiceStatusTab />;
       case 'mcp':
         return <McpTab />;
-      case 'bug-report':
-        return <BugReportTab />;
+
       default:
         return null;
     }
